+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Added
 - Show parsed AST and tokens debugging tasks/commands/menus generated for each language.
+- Show analyzed AST debugging tasks/commands/menus generated for each language.
 
 ### Changed
 - Restructure troubleshooting documentation.